pyproject.toml
[tool.poetry]
name = "django-comments-tree"
version = "0.1.9"
description = "Django Comments Framework extension app with django-treebeard tree support, follow up notifications and email confirmations."
authors = ["Ed Henderson <ed@sharpertool.com>", "Alexander @PekopT Brilliantov <alex@brilliantov.me>"]

[tool.poetry.dependencies]
python = "^3.4.0"
django-contrib-comments = "^1.8.0"
djangorestframework = "^3.6.0"
six = "^1.12.0"
Django = "^2.0.0"
docutils = "^0.14.0"
django-markdown2 = "^0.3.1"
django-markupfield = "^1.5.1"

[tool.poetry.dev-dependencies]
coverage = "^4.5.0"
mock = "^2.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
